Reorganize migrated packages
