+++ pkg/controller/machine/testactuator.go
@@ -21,7 +21,7 @@ import (
 	"sync"
 
 	"github.com/openshift/cluster-api/pkg/apis/cluster/v1alpha1"
-	"github.com/openshift/cluster-api/pkg/apis/machine/v1beta1"
+	"github.com/openshift/machine-api-operator/pkg/apis/machine/v1beta1"
 )
 
 var _ Actuator = &TestActuator{}
