+++ pkg/controller/machineset/delete_policy.go
@@ -20,7 +20,7 @@ import (
 	"math"
 	"sort"
 
-	"github.com/openshift/cluster-api/pkg/apis/machine/v1beta1"
+	"github.com/openshift/machine-api-operator/pkg/apis/machine/v1beta1"
 	"github.com/pkg/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
