+++ pkg/controller/machineset/machine_test.go
@@ -19,7 +19,7 @@ package machineset
 import (
 	"testing"
 
-	"github.com/openshift/cluster-api/pkg/apis/machine/v1beta1"
+	"github.com/openshift/machine-api-operator/pkg/apis/machine/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
